[Dev] Raspberry pi headless setup

+++ raspberrypi/headless/firstboot.sh
@@ -1,4 +1,4 @@
-#!/bin/bash
+#!/usr/bin/env bash
 
 # -----------------------------------------------------------------------------
 # Hostname
@@ -22,21 +22,25 @@ systemctl restart avahi-daemon
 username="tedi"
 
 # Remove pi user
-pkill -u pi
-passwd --lock pi
-deluser --remove-home pi
+if id pi &>/dev/null; then
+	pkill -u pi
+	passwd --lock pi
+	deluser --remove-home pi
+fi
 
 # Add new user
-adduser --disabled-password --gecos=",,,," "${username}"
-passwd -d "${username}"
-usermod -a -G adm,dialout,cdrom,sudo,audio,video,plugdev,games,users,input,netdev,gpio,i2c,spi "${username}"
+if ! id "${username}" &>/dev/null; then
+	adduser --disabled-password --gecos=",,,," "${username}"
+	passwd -d "${username}"
+	usermod -a -G adm,dialout,cdrom,sudo,audio,video,plugdev,games,users,input,netdev,gpio,i2c,spi "${username}"
+fi
 
 # -----------------------------------------------------------------------------
 # Security
 
 # Ssh public key
 mkdir -p /home/"${username}"/.ssh/
-mv /boot/firstboot/id_rsa_init.pub /home/"${username}"/.ssh/ || exit 1
+cp /boot/firstboot/id_rsa_init.pub /home/"${username}"/.ssh/ || exit 1
 chown "${username}":"${username}" /home/"${username}"/.ssh/
 chmod 700 /home/"${username}"/.ssh/
 chown "${username}":"${username}" /home/"${username}"/.ssh/id_rsa_init.pub
@@ -51,7 +55,20 @@ chmod 600 /home/"${username}"/.ssh/authorized_keys
 # Packages
 
 # Update apt cache, ingore "Release file... is not valid yet." error
-# apt-get update -o Acquire::Check-Valid-Until=false -o Acquire::Check-Date=false
+apt-get update -o Acquire::Check-Valid-Until=false -o Acquire::Check-Date=false
 
-# Install git client
-# apt-get install -y git
+# Install packages
+apt-get install -y ufw
+
+# -----------------------------------------------------------------------------
+# Firewall
+
+command -v ufw > /dev/null 2>&1 &&
+	{
+		# Deny everything except ssh
+		ufw default deny incoming
+		# ufw default deny outgoing
+		# ufw allow proto tcp from any to any port 80,443
+		ufw allow 22/tcp
+		ufw --force enable
+	} || exit 1
